refactor: add oop
