+++ algorithms/sorting/quick_sort.py
@@ -1,18 +1,7 @@
-from typing import Generic, TypeVar, Callable
-from operator import lt
-
-T = TypeVar('T')
-
-class QuickSort(Generic[T]):
-    def __init__(self, lst: list[T], comp_func: Callable[[T, T], bool] = lt) -> None:
-        self.comp_func = comp_func
-        self.lst = lst
+from algorithms.sorting.sort import Sort, T
 
 
-    def _swap(self, first_index: int, second_index: int) -> None:
-        self.lst[first_index], self.lst[second_index] = self.lst[second_index], self.lst[first_index]
-
-    
+class QuickSort(Sort):
     def _get_left_right(self, left, right, pivot):
         while left < right and (self.comp_func(self.lst[left], self.lst[pivot]) or self.lst[left] == self.lst[pivot]):
             left += 1
